# libs/emoji-tracker-lib/src/api/openapi/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"AddUserReqBody": {
				"properties": {
					"email": {
						"type": "string"
					},
					"username": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"username"
				],
				"type": "object"
			},
			"AddMoodReqBody": {
				"properties": {
					"note": {
						"type": "string"
					},
					"mood": {
						"type": "string"
					}
				},
				"required": [
					"mood"
				],
				"type": "object"
			},
			"MoodStats": {
				"properties": {
					"note": {
						"type": "string"
					},
					"moodCounts": {
						"type": "string"
					},
					"emoji": {
						"type": "string"
					}
				},
				"required": [
					"moodCounts",
					"emoji"
				],
				"type": "object"
			},
			"SuccessResponse_MoodStats-Array_": {
				"properties": {
					"response": {
						"items": {
							"$ref": "#/components/schemas/MoodStats"
						},
						"type": "array"
					}
				},
				"required": [
					"response"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponse_any_": {
				"properties": {
					"response": {}
				},
				"required": [
					"response"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponse_string_": {
				"properties": {
					"response": {
						"type": "string"
					}
				},
				"required": [
					"response"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "emoji-tracker",
		"version": "v1.0.0",
		"license": {
			"name": "emojiTracker"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/users/create": {
			"post": {
				"operationId": "HandlerFunc",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"summary": "create new user",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddUserReqBody"
							}
						}
					}
				}
			}
		},
		"/users/{userId}/mood": {
			"post": {
				"operationId": "HandlerFunc",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"summary": "create new user's mood",
				"tags": [
					"Moods"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddMoodReqBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "HandlerFunc",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"summary": "delete user's mood",
				"tags": [
					"Moods"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "moodId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"get": {
				"operationId": "HandlerFunc",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse_MoodStats-Array_"
								}
							}
						}
					}
				},
				"summary": "get user's mood\nfilter with this-month, previous-month, last-month",
				"tags": [
					"Moods"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "filter",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "HandlerFunc",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"summary": "update user's mood",
				"tags": [
					"Moods"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "moodId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddMoodReqBody"
							}
						}
					}
				}
			}
		},
		"/users/share/{hash}": {
			"get": {
				"operationId": "HandlerFunc",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse_any_"
								}
							}
						}
					}
				},
				"summary": "get user's mood\nfilter with this-month, previous-month, last-month",
				"tags": [
					"Sharing"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "hash",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/{userId}/share/mood": {
			"get": {
				"operationId": "HandlerFunc",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse_string_"
								}
							}
						}
					}
				},
				"summary": "get user's mood\nfilter with this-month, previous-month, last-month",
				"tags": [
					"Sharing"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "filter",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:3334/v1",
			"description": "Local Env"
		}
	]
}